move deploy.ts and format code
