+++ packages/sc-toolkit/commands/init/assembly/contracts/run.ts
